sercomm_cons: Utilize msgb's better, don't always end at \n
In the sercomm_cons layer, ee used to enqueue a msgb for sending
every time there is an end-of-line.  However, if we send a number
of very short lines, we easily run out of msgbs.

Now we check how much msgb backlog there is in the transmit queue,
and decide to skip the end-of-line flushing if needed.

This still doesn't solve all our problems, but its still a useful
mechanism.