TWL3025: Ensure we don't perform downlink calibration too early
If we perform the downlink calibration too early, the TRF6151 might
not yet provide a stable signal and we'll not be able to receive
anything.

From the desired "BDLENA" time, we subtract all the delays and
latencies to determine the point in time at which the calibration
process should start.