stdint.h: Make it work with my GCC 4.3.3/GLIBC 2.9 toolchain
The compiler complains about seing two conflicting definitions
for int32_t and the compile will stop. One definitions comes from
here and the other is coming from sys/types.h.

One option would be to try to use unsigned int and signed int
instead of using a long and see if that is matching, the other
option is to include the real libc stdint.h header and see if
it has defined a uint32_t for us.