layer1/mframe_sched: Make task update safer
We try to prevent switching task at a bad time. The algorithm
just tries to find the first fn at which it's safe to activate
a previously disabled task.

tasks     -> Running tasks
tasks_tgt -> What we aim to run

Since tasks_tgt is still only read, it's safe for the l1a to
modify it without locking.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>