[layer1] Power Measurement updates
Instead of using the 'normal burst' receive window, use the proper
RXWIN_PW for power measurements. However, in this case, one PW window
corresponds to one PM measurement and in order to do two measurements,
multiple PW windows need to be set-up in the TPU.  Since we don't support
that yet, we simply switch to doing only one power measurement for now.