Allow top-level Makefile to use arm-none-eabi toolchain
The gnuarm.com toolchain works fine but is very old. And although
it is based on newlib, the tuple that we used to configure for was
arm-elf-linux, which is bogus since we aren't building for Linux.

This patch optimizes for the CodeSourcery G++ Lite 2010.09 ARM EABI
toolchain instead, and for libosmocore installation it assumes that
the new toolchain was unpacked next to the old one. Download it here:

http://www.codesourcery.com/sgpp/lite/arm/portal/release1592

(The Linux Installer seems not to work reliably so I recommend the
 TAR.)

Since CC is detected by configure when the host tuple points to a sane
toolchain we shouldn't hard-code the gnuarm.com compiler.

The patch autodetects arm-elf-gcc installed in PATH, and uses arm-elf
as prefix if it is found. Otherwise, it defaults to arm-none-eabi.

make CROSS_HOST=arm-xyzzy can be used to override on the command line.

Author: Peter Stuge <peter@stuge.se>
Acked-by: Sylvain Munaut <tnt@246tNt.com>