firmware/monitor: Improvements and sync support
By pressing the green button, the sync mode is entered. The screen
show some informations about the cell. The beep indicates, if the
received BCCH was valid or not.

By pressing the Down button, the list of channels of the serving cell
and neighbour cells can be viewed and scrolled through.

By pressing the Up button, the levels (downlink or uplink) of the serving
cell can be viewed. Also it is possible to select one of the serving
cell's frequencies by pressing the Left/Right button.