debian: Testing was disabled because of a funny x86 compiler issue caused by us
The bits test will fail on Debian Stable for x86. Somehow the load
of a static const array is ending up accessing random memory. This
only happens when not using any -OX options and this is caused by
us. Do not override the default debian CFLAGS.. and the tests will
pass on debian stable.