mobile: Add user-user message transmission to call control
If a user-user message is received, it is displayed via VTY.

If a user-user message is defined with "call <ms> set-useruser <text>",
it will can be transmitted with "call <ms> useruser" while there is an
active call. After transmission, the message is cleared. If defined, it
will be transmitted with next setup/alerting/connect/disconnect/release
message also.