Add gain control to calypso BTS code
The gain of RX does not follow the sync-cell anymore. It still uses the
appropiate gain for sync-cell, but for RX of BTS bursts it uses a gain
that is suitable for a receive level of -47 dBm (by default). This is a
gain level of 0.

The low gain of 0 is the best start value, if phones are close to
the BTS, but also for low RX levels below -100 dBm, as several tests
showed.