Always forward the strongest received access burst of one multiframe
Instead of forwarding access bursts that are above some average noise
floor, the strongest of burst of a multiframe is always forwarded, no
matter how weak it is.

Since bandwidth of serial interface is limited, it is not possible to
forward all received access bursts. If there are suddenly many access
bursts or if a temporary noise causes detection of several bursts, only
the strongest of them is forwarded. This prevents overflow of serial
interface.

Detection of bursts is possible, no matter how weak they are. Tests have
shown that detection works up to 100%, even at border of BTS coverage.